==> notebooks/1_Download_new.ipynb <==
!pip3 install kaggle

import pandas as pd
import numpy as np
import os,time
from tqdm import tqdm
import zipfile
import glob

data_dir = "../data/"

#!kaggle competitions download -c otto-recommender-system -p ../data/
#!unzip -q "../data/otto-recommender-system.zip" -d ../data/

with zipfile.ZipFile(f"{data_dir}otto-recommender-system.zip", 'r') as zip_ref:
    zip_ref.extractall(data_dir)

id2type = ['clicks','carts','orders']
type2id = {a:i for i,a in enumerate(id2type)}
print(id2type, type2id)

chunksize = 100_000
#from the given json file, below function converts to multiple parquet files of 100,000 rows each, 
#enabling better memory management and easier handling with pandas
def jsontodf(fname,total_chunks,prefix_name,output_directory):
    sessions = []
    aids = []
    tss = []
    types =[]
    chunks = pd.read_json(fname,lines=True,chunksize=100_000)
    for e, chunk in enumerate(tqdm(chunks, total=total_chunks)):
        event_dict = {
            'session': [],
            'aid': [],
            'ts': [],
            'type': [],
        }

        for session, events in zip(chunk['session'].tolist(), chunk['events'].tolist()):
            for event in events:
                event_dict['session'].append(session)
                event_dict['aid'].append(event['aid'])
                event_dict['ts'].append(event['ts']//1000)
                event_dict['type'].append(type2id[event['type']])

        # save DataFrame
        start = str(e*chunksize).zfill(9)
        end = str(e*chunksize+chunksize).zfill(9)
        temp = pd.DataFrame(event_dict)
        temp["type"] = temp['type'].astype(np.uint8)
        temp.ts = temp.ts.astype('int32')
        temp.to_parquet(os.path.join(output_directory,f"{prefix_name}_{start}_{end}.parquet"))

jsontodf(os.path.join(data_dir,'train.jsonl'),129,"train",os.path.join(data_dir,"train_pqt_chunks"))

jsontodf(os.path.join(data_dir,'test.jsonl'),17,"test",os.path.join(data_dir,"test_pqt_chunks"))

## File for CrossValidation
we would pick last 7 days as a validation data and use 3 weeks to train and the last week to validation.

def load_train_data():    
    dfs = []
    for e, chunk_file in enumerate(glob.glob(f'{data_dir}train_pqt_chunks/*')):
        chunk = pd.read_parquet(chunk_file)
        dfs.append(chunk)
        
    train_df = pd.concat(dfs).reset_index(drop=True) #.astype({"ts": "datetime64[ms]"})
    return train_df

seven_days = 7*24*60*60
train = load_train_data()
train_cutoff = train.ts.max() - seven_days
valid = train[train.ts > train_cutoff]
train = train[train.ts <= train_cutoff]

train_cutoff = train.ts.max() - seven_days
train_3rdweek = train[train.ts > train_cutoff]
train_f2weeks = train[train.ts <= train_cutoff]

overlapping_sessions = set(train_f2weeks.session).intersection(set(train_3rdweek.session))
train_3rdweek = train_3rdweek[~train_3rdweek.session.isin(overlapping_sessions)]

print(f"train_3rdweek starting from {time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(train_3rdweek.ts.min()))} and ending at {time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(train_3rdweek.ts.max()))}")
print(f"validation data starting from {time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(valid.ts.min()))} and ending at {time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(valid.ts.max()))}")
print(f"sessions between the 2 datasets = {len(set(train_f2weeks.session).intersection(set(train_3rdweek.session)))}")

train_3rdweek.to_parquet(f"{data_dir}local_cv/3week_train_data.parquet")

ground_truth = valid.groupby(['session', 'type'])['aid'].apply(list)
ground_truth = ground_truth.reset_index().rename(columns={'aid': 'labels'})
ground_truth.to_parquet(f"{data_dir}/local_cv/4week_train_labels.parquet")
