==> notebooks/4_Candidate_gen_Ranking.ipynb <==
!pip3 install gensim polars annoy hyperopt

import pandas as pd
import numpy as np
import glob
import itertools
from collections import Counter
import polars as pl
from gensim.models import Word2Vec
from annoy import AnnoyIndex
import random
from sklearn.model_selection import GroupKFold
from hyperopt import STATUS_OK, Trials, fmin, hp, tpe

DISK_PIECES = 4
data_dir = "../data/"
type_labels = {'clicks':0, 'carts':1, 'orders':2}
type_weight = {0:1, 1:6, 2:3}

def load_data(path):    
    dfs = []
    for e, chunk_file in enumerate(glob.glob(path)):
        chunk = pd.read_parquet(chunk_file)
        dfs.append(chunk)
    return pd.concat(dfs).reset_index(drop=True) #.astype({"ts": "datetime64[ms]"})


train_df = load_data(f'{data_dir}/train_pqt_chunks/*')
test_df = load_data(f'{data_dir}/test_pqt_chunks/*')

print('shape',train_df.shape, test_df.shape)

sentences_df = pl.from_pandas(pd.concat([train_df,test_df])).groupby('session').agg(pl.col('aid').alias('sentence'))
sentences = sentences_df['sentence'].to_list()
w2vec = Word2Vec(sentences=sentences, vector_size=32, min_count=1, workers=6)
w2vec.save("../models/otto_word2vec.model")
del sentences_df, sentences, w2vec

w2vec = Word2Vec.load("../models/otto_word2vec.model")
embedding_weight = np.load('../models/otto_word2vec.model.wv.vectors.npy')
embedding_weight_neg = np.load('../models/otto_word2vec.model.syn1neg.npy')
embedding_weigh_dict_df = pl.from_pandas(pd.DataFrame(embedding_weight, columns = ['Embedding_' + str(x) for x in range(32)]).reset_index().rename(columns = {'index':'aid'})).with_columns(pl.col('aid').cast(pl.Int32))


aid2idx = {aid: i for i, aid in enumerate(w2vec.wv.index_to_key)}
index = AnnoyIndex(32, 'euclidean')

for aid, idx in aid2idx.items():
    index.add_item(idx, w2vec.wv.vectors[idx])
    
index.build(10)

def get_nn(aid,num_candidates):
    return [w2vec.wv.index_to_key[i] for i in index.get_nns_by_item(aid2idx[aid], num_candidates+1)[1:]]

def pqt_to_dict(df):
    return df.groupby('aid_x').aid_y.apply(list).to_dict()

# LOAD THREE CO-VISITATION MATRICES
top_40_clicks = pqt_to_dict( pd.read_parquet(f'{data_dir}top_40_clicks_0.pqt') )
for k in range(1,DISK_PIECES): 
    top_40_clicks.update( pqt_to_dict( pd.read_parquet(f'{data_dir}top_40_clicks_{k}.pqt') ) )

top_40_buys = pqt_to_dict( pd.read_parquet(f'{data_dir}top_40_carts_orders_0.pqt') )
for k in range(1,DISK_PIECES): 
    top_40_buys.update( pqt_to_dict( pd.read_parquet(f'{data_dir}top_40_carts_orders_{k}.pqt') ) )
top_40_buy2buy = pqt_to_dict( pd.read_parquet(f'{data_dir}top_40_buy2buy.pqt') )

# TOP CLICKS AND ORDERS IN TEST
top_clicks = test_df.loc[test_df['type']=='clicks','aid'].value_counts().index.values[:40]
top_orders = test_df.loc[test_df['type']=='orders','aid'].value_counts().index.values[:40]

print('Here are size of our 3 co-visitation matrices:')
print( len( top_40_clicks ), len( top_40_buy2buy ), len( top_40_buys ) )

type_weight_multipliers = {0: 3, 1: 6, 2: 3}
def suggest_clicks(df,num_candidates):
    # USER HISTORY AIDS AND TYPES
    aids=df.aid.tolist()
    types = df.type.tolist()
    unique_aids = list(dict.fromkeys(aids[::-1] ))
    # RERANK CANDIDATES USING WEIGHTS
    if len(unique_aids)>=num_candidates:
        weights=np.logspace(0.1,1,len(aids),base=2, endpoint=True)-1
        aids_temp = Counter() 
        # RERANK BASED ON REPEAT ITEMS AND TYPE OF ITEMS
        for aid,w,t in zip(aids,weights,types): 
            aids_temp[aid] += w * type_weight_multipliers[t]
        sorted_aids = [k for k,v in aids_temp.most_common(num_candidates)]
        return sorted_aids
    # USE "CLICKS" CO-VISITATION MATRIX
    aids2 = list(itertools.chain(*[top_40_clicks[aid] for aid in unique_aids if aid in top_40_clicks]))
    #WORD2VEC CANDIDATES based on last 5 aids
    recent_aids=aids[::-1]
    word2vec_candidates = list(itertools.chain(*[get_nn(recent_aids[i],num_candidates) for i in range(min(5,len(recent_aids)))]))
    #Matrix Factorization based candidates
    #matrix_fact_candidates = list(itertools.chain(*[aid_matrix_fact[recent_aids[i]] for i in range(min(5,len(recent_aids)))]))
    # RERANK CANDIDATES
    #top_aids2 = [aid2 for aid2, cnt in Counter(aids2+word2vec_candidates+matrix_fact_candidates).most_common(num_candidates) if aid2 not in unique_aids]
    top_aids2 = [aid2 for aid2, cnt in Counter(aids2+word2vec_candidates).most_common(num_candidates) if aid2 not in unique_aids]
    
    result = unique_aids + top_aids2[:num_candidates - len(unique_aids)]
    # USE TOP20 TEST CLICKS
    return result + list(top_clicks)[:num_candidates-len(result)]

def suggest_buys(df,num_candidates):
    # USER HISTORY AIDS AND TYPES
    aids=df.aid.tolist()
    types = df.type.tolist()
    # UNIQUE AIDS AND UNIQUE BUYS
    unique_aids = list(dict.fromkeys(aids[::-1] ))
    df = df.loc[(df['type']==1)|(df['type']==2)]
    unique_buys = list(dict.fromkeys( df.aid.tolist()[::-1] ))
    # RERANK CANDIDATES USING WEIGHTS
    if len(unique_aids)>=num_candidates:
        weights=np.logspace(0.5,1,len(aids),base=2, endpoint=True)-1
        aids_temp = Counter() 
        # RERANK BASED ON REPEAT ITEMS AND TYPE OF ITEMS
        for aid,w,t in zip(aids,weights,types): 
            aids_temp[aid] += w * type_weight_multipliers[t]
        # RERANK CANDIDATES USING "BUY2BUY" CO-VISITATION MATRIX
        aids3 = list(itertools.chain(*[top_20_buy2buy[aid] for aid in unique_buys if aid in top_20_buy2buy]))
        for aid in aids3: aids_temp[aid] += 0.1
        sorted_aids = [k for k,v in aids_temp.most_common(num_candidates)]
        return sorted_aids
    # USE "CART ORDER" CO-VISITATION MATRIX
    aids2 = list(itertools.chain(*[top_20_buys[aid] for aid in unique_aids if aid in top_20_buys]))
    # USE "BUY2BUY" CO-VISITATION MATRIX
    aids3 = list(itertools.chain(*[top_20_buy2buy[aid] for aid in unique_buys if aid in top_20_buy2buy]))
    # USE "Word2vec" Matrix
    recent_aids=list(dict.fromkeys( df.loc[(df['type']==1)].aid.tolist()[::-1] ))
    word2vec_candidates = list(itertools.chain(*[get_nn(recent_aids[i],num_candidates) for i in range(min(3,len(recent_aids)))]))
    # RERANK CANDIDATES
    top_aids2 = [aid2 for aid2, cnt in Counter(aids2+aids3+word2vec_candidates).most_common(num_candidates) if aid2 not in unique_aids] 
    result = unique_aids + top_aids2[:num_candidates - len(unique_aids)]
    # USE TOP20 TEST ORDERS
    return result + list(top_orders)[:num_candidates-len(result)]


def click_recall_score(pred_df,test_labels):
    score = 0
    weights = {'clicks': 0.10, 'carts': 0.30, 'orders': 0.60}
    for t in ['clicks']:
        sub = pred_df.loc[pred_df.session_type.str.contains(t)].copy()
        sub['session'] = sub.session_type.apply(lambda x: int(x.split('_')[0]))
        sub.labels = sub.labels.apply(lambda x: [int(i) for i in x.split(' ')[:20]])
        test_labels = test_labels.loc[test_labels['type']==t]
        test_labels = test_labels.merge(sub, how='left', on=['session'])
        
        test_labels['hits'] = test_labels.apply(lambda df: len(set(df.ground_truth).intersection(set(df.labels))), axis=1)
        test_labels['gt_count'] = test_labels.ground_truth.str.len().clip(0,20)
        recall = test_labels['hits'].sum() / test_labels['gt_count'].sum()
        score += weights[t]*recall
        print(f'{t} recall =',recall)

train_df = pd.read_parquet(f"{data_dir}local_cv/3week_train_data.parquet")
train_df = train_df[train_df.type==0]


test_labels = pd.read_parquet("../data/local_cv/4week_train_labels.parquet")
test_labels.head()

pred_df_clicks = train_df.sort_values(["session", "ts"]).groupby(["session"]).apply(
    lambda x: suggest_clicks(x,20)
)

del train_df, top_40_clicks, w2vec, embedding_weight, embedding_weight_neg, embedding_weigh_dict_df, index

train_df_click_recommended = pl.from_pandas(pd.DataFrame(pred_df_clicks, columns = ['aid']).reset_index()).explode("aid")
train_df_click_recommended = train_df_click_recommended.with_columns([
    pl.lit(0).alias('ts').cast(pl.Int32),    
    pl.lit(0).alias('type').cast(pl.Int8),
    pl.col('session').cast(pl.Int32),
    pl.col('aid').cast(pl.Int32),
])

n_col_after_join = train_df_click_recommended.groupby('session').agg([
    pl.col('aid').cumcount().alias('CG_ranking')]).select(
    pl.col('CG_ranking').explode().cast(pl.Int64))
train_df_click_recommended = pl.concat([train_df_click_recommended, n_col_after_join], how="horizontal")

train_df_click_recommended.head()


# calculate recall for ranking
pred_df = pd.DataFrame(pred_df_clicks.add_suffix("_clicks"), columns=["labels"]).reset_index()
clicks_pred_df = pd.DataFrame(pred_df_clicks.add_suffix("_clicks"), columns=["labels"]).reset_index()
pred_df.columns = ["session_type", "labels"]
pred_df["labels"] = pred_df["labels"].map(lambda x:x[0:20])
pred_df["labels"] = pred_df.labels.apply(lambda x: " ".join(map(str,x)))

del pred_df_clicks
pred_df.head()

aids = test_labels.labels.explode().astype('int32').rename('item')
tar = test_labels[['session']].astype('int32').rename({'session':'user'},axis=1)
tar = tar.merge(aids, left_index=True, right_index=True, how='left')
tar['click'] = 1
tar.columns = ["session","aid","click"]
tar = pl.from_pandas(tar).with_columns([pl.col('session').cast(pl.Int32),pl.col('aid').cast(pl.Int32)])
train_df_click_recommended = train_df_click_recommended.join(tar,on=['session','aid'],how='left').with_columns(pl.col("click").fill_null(pl.lit(0),),)
del tar, aids

train_df_click_recommended = (train_df_click_recommended.filter(pl.col("click") == 0).sample(frac=0.5)
                                .vstack(train_df_click_recommended.filter(pl.col("click") == 1)))

aid_features = pl.read_parquet("../data/features/item_feat.pqt").with_columns(pl.col('aid').cast(pl.Int32))
train_df_click_recommended = train_df_click_recommended.join(aid_features, how='left',on="aid")
del aid_features

user_features = pl.read_parquet("../data/features/user_feat.pqt").with_columns(pl.col('session').cast(pl.Int32))
train_df_click_recommended = train_df_click_recommended.join(user_features,how='left', on='session')
del user_features

user_item_features = pl.read_parquet("../data/features/user_item.pqt").with_columns([pl.col('session').cast(pl.Int32),pl.col('aid').cast(pl.Int32)])
train_df_click_recommended = train_df_click_recommended.join(user_item_features,how='left', on=['session','aid'])
del user_item_features

train_df_click_recommended = train_df_click_recommended.fill_null(-1) 

train_df_click_recommended = train_df_click_recommended.select(pl.exclude(["__index_level_0__","__index_level_0___right","ts_right","type_right"]))

train_df_click_recommended.head()

train_df = pd.read_parquet(f"{data_dir}local_cv/3week_train_data.parquet")
train_df = train_df[train_df.type==0]
real_action = pl.from_pandas(train_df[["session","aid"]]).with_columns([pl.lit(1).alias('real_action'),pl.col('session').cast(pl.Int32),pl.col('aid').cast(pl.Int32)])
del train_df
train_df_click_recommended = train_df_click_recommended.join(real_action,how='left', on=['session','aid']).with_columns(pl.col("real_action").fill_null(pl.lit(0),),)
del real_action

candidates = train_df_click_recommended.to_pandas()
del train_df_click_recommended
#candidates.loc[candidates.aid.isin(top_clicks_l3days),"is_top_clicks_3days"] = 1
#candidates.loc[candidates.aid.isin(top_clicks_l1days),"is_top_clicks_1days"] = 1
#candidates.loc[candidates.aid.isin(top_clicks),"is_top_clicks_7days"] = 1
#candidates.fillna(value = {"is_top_clicks_3days": 0, "is_top_clicks_1days": 0, "is_top_clicks_7days": 0},inplace=True)
candidates.head()

candidates.click = candidates.click.astype(int)
candidates["click"].value_counts()

test_sessions = random.choices(candidates.session.unique(), k=candidates.session.nunique()//2)
test_candidates = candidates[candidates.session.isin(test_sessions)]
train_candidates = candidates[~candidates.session.isin(test_sessions)]
print(train_candidates.shape,test_candidates.shape)

del candidates

FEATURES = [col for col in train_candidates.columns.values.tolist() if col not in ["session","aid","click","ts","type"]]
FEATURES

space={'max_depth': hp.quniform("max_depth", 3, 12, 1),
        'colsample_bytree' : hp.uniform('colsample_bytree', 0.5,1),
        'min_child_weight' : hp.quniform('min_child_weight', 1000, 100000, 1000),
        'num_boost_round': hp.quniform('num_boost_round', 50,1000,50),
        'subsample': hp.uniform('learning_rate', 0.4,1),
        'objective':hp.choice('objective',['rank:pairwise','rank:ndcg','rank:map']),
        'seed': 0}

X_train = train_candidates.loc[:, FEATURES]
y_train = train_candidates.loc[:, 'click']
X_valid = test_candidates.loc[:, FEATURES]
y_valid = test_candidates.loc[:, 'click']

test_labels = test_labels[test_labels['type']=="clicks"]
test_labels = test_labels[test_labels.session.isin(test_candidates["session"].unique())]


# IF YOU HAVE 50 CANDIDATE WE USE 50 BELOW
dtrain = xgb.DMatrix(X_train, y_train, group=train_candidates[["session"]].groupby("session")['session'].count().values) 
dvalid = xgb.DMatrix(X_valid, y_valid, group=test_candidates[["session"]].groupby("session")['session'].count().values) 
def objective(space):
    clf=xgb.train(params={'max_depth':int(space['max_depth']),
                    'min_child_weight':int(space['min_child_weight']), 
                   'colsample_bytree':space['colsample_bytree'], 
                   'subsample' : space['subsample'],
                    'tree_method': "auto",
                    'objective' : space['objective']},
                    num_boost_round = int(space['num_boost_round']),
                      dtrain=dtrain,verbose_eval=100
                        )
    
    preds = np.zeros(len(test_candidates))
    predictions = test_candidates[['session','aid']].copy()
    dtest = xgb.DMatrix(data=X_valid)
    predictions['pred'] = clf.predict(dvalid)

    predictions = predictions.sort_values(['session','pred'], ascending=[True,False]).reset_index(drop=True)
    predictions['n'] = predictions.groupby('aid').aid.cumcount().astype('int8')
    predictions = predictions.loc[predictions.n<20]
    sub = predictions.groupby('session').aid.apply(list)
    sub = sub.to_frame().reset_index()
    sub.aid = sub.aid.apply(lambda x: " ".join(map(str,x)))
    sub.columns = ['session_type','labels']
    
    sub.session_type = sub.session_type.astype('str')+ '_clicks'
    sub = sub.loc[sub.session_type.str.contains('clicks')].copy()
    sub['session'] = sub.session_type.apply(lambda x: int(x.split('_')[0]))
    sub.labels = sub.labels.apply(lambda x: [int(i) for i in x.split(' ')[:20]])
    verify = test_labels.merge(sub, how='left', on=['session'])
    verify.dropna(inplace=True)
    verify['hits'] = verify.apply(lambda df: len(set(df.ground_truth).intersection(set(df.labels))), axis=1)
    verify['gt_count'] = verify.ground_truth.str.len().clip(0,20)
    recall = verify['hits'].sum() / verify['gt_count'].sum()
    return {'loss': -recall, 'status': STATUS_OK }

trials = Trials()

best_hyperparams = fmin(fn = objective,
                        space = space,
                        algo = tpe.suggest,
                        max_evals = 1,
                        trials = trials)

print(best_hyperparams)
